casi

+++ srcs/init_forks.c
@@ -6,7 +6,7 @@
 /*   By: adpachec <adpachec@student.42madrid.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/09 11:41:38 by adpachec          #+#    #+#             */
-/*   Updated: 2023/05/10 11:42:15 by adpachec         ###   ########.fr       */
+/*   Updated: 2023/05/11 14:21:21 by adpachec         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,10 +24,10 @@ static t_fork	*init_fork(int id)
 	return (fork);
 }
 
-t_fork	**init_forks(int num_forks)
+t_fork	**init_forks(unsigned long num_forks)
 {
-	int		i;
-	t_fork	**forks;
+	unsigned long	i;
+	t_fork			**forks;
 
 	forks = (t_fork **)ft_calloc(sizeof(t_fork *), num_forks + 1);
 	if (!forks)
